--- a/config/plugins/fugitive.nix
+++ b/config/plugins/fugitive.nix
@@ -15,6 +15,17 @@ _: {
     # unstoppable git plugin
     # https://github.com/tpope/vim-fugitive/
     # https://nix-community.github.io/nixvim/plugins/fugitive
-    plugins.fugitive.enable = true;
+    plugins.fugitive = {
+      enable = true;
+      # lazyLoad.enable = true;
+      # TODO: not working... I think maybe only works for mkNeovimPlugin?
+      # lazyLoad = {
+      #   settings = {
+      #     cmd = "Git";
+      #     keys = [ "<leader>g" ];
+      #     ft = "fugitive";
+      #   };
+      # };
+    };
   };
 }