--- a/config/plugins/lsp-lines.nix
+++ b/config/plugins/lsp-lines.nix
@@ -5,6 +5,7 @@ _: {
   config = {
     plugins.lsp-lines = {
       enable = true;
+      lazyLoad.settings.event = "BufEnter";
       # currentLine = true;
     };
   };