--- a/config/plugins/luasnip.nix
+++ b/config/plugins/luasnip.nix
@@ -1,4 +1,5 @@
-_: {
+{ config, ... }:
+{
   # useful code expansions
   # https://github.com/L3MON4D3/LuaSnip
   # https://nix-community.github.io/nixvim/plugins/luasnip
@@ -6,6 +7,7 @@ _: {
     plugins.luasnip = {
       enable = true;
       lazyLoad.settings.event = "BufEnter";
+      lazyLoad.enable = config.lazyLoad.enable;
       settings = {
         #   enable_autosnippets = true;
         #   store_selection_keys = “<Tab>”;