--- a/config/plugins/luasnip.nix
+++ b/config/plugins/luasnip.nix
@@ -5,6 +5,7 @@ _: {
   config = {
     plugins.luasnip = {
       enable = true;
+      lazyLoad.settings.event = "BufEnter";
       settings = {
         #   enable_autosnippets = true;
         #   store_selection_keys = “<Tab>”;