--- a/config/plugins/nvim-colorizer.nix
+++ b/config/plugins/nvim-colorizer.nix
@@ -5,6 +5,7 @@ _: {
   config = {
     plugins.colorizer = {
       enable = true;
+      lazyLoad.settings.event = "BufEnter";
       settings = {
         user_default_options = {
           RGB = true;