--- a/config/plugins/oil-git-status.nix
+++ b/config/plugins/oil-git-status.nix
@@ -3,7 +3,11 @@ let
   oil-git-status = pkgs.vimUtils.buildVimPlugin {
     name = "oil-git-status.nvim";
     src = inputs.oil-git-status;
-    dependencies = [ pkgs.vimPlugins.oil-nvim ];
+    dependencies = with pkgs.vimPlugins; [
+      oil-nvim
+      # TODO: a better way?
+      lz-n
+    ];
   };
 in
 {
@@ -12,8 +16,17 @@ in
   config = {
     extraConfigLua = # lua
       ''
-        require('oil-git-status').setup()
+        require("lz.n").load({
+          {
+            "oil-git-status.nvim",
+            after = function()
+              require("oil-git-status").setup()
+            end,
+            ft = "oil",
+          },
+        })
       '';
+
     extraPlugins = [ oil-git-status ];
   };
 }