--- a/config/plugins/oil.nix
+++ b/config/plugins/oil.nix
@@ -1,4 +1,5 @@
-_: {
+{ config, lib, ... }:
+{
   # easily browse directories
   # https://github.com/stevearc/oil.nvim/
   # https://nix-community.github.io/nixvim/plugins/oil
@@ -14,6 +15,7 @@ _: {
           ft = "netrw";
         };
       };
+      lazyLoad.enable = config.lazyLoad.enable;
       settings = {
         view_options = {
           show_hidden = true;
@@ -24,36 +26,37 @@ _: {
       };
     };
 
-    extraConfigLua = # lua
-      ''
-        -- Helper function to check if a path is a directory (handling oil:// prefix)
-        local function remove_oil_prefix(path)
-          return vim.startswith(path, "oil://") and path:sub(7) or path
-        end
-        local function is_directory(path)
-          return vim.fn.isdirectory(remove_oil_prefix(path)) == 1
-        end
+    extraConfigLua =
+      lib.mkIf config.plugins.oil.lazyLoad.enable
+        # lua
+        ''
+          local function remove_oil_prefix(path)
+            return vim.startswith(path, "oil://") and path:sub(7) or path
+          end
+          local function is_directory(path)
+            return vim.fn.isdirectory(remove_oil_prefix(path)) == 1
+          end
 
-        -- Handle `nvim somedir`
-        vim.api.nvim_create_autocmd("VimEnter", {
-          callback = function(data)
-            if is_directory(data.file) then
-              vim.schedule(function()
-                require("oil").open(data.file)
-              end)
-            end
-          end,
-        })
+          -- Handle `nvim somedir`
+          vim.api.nvim_create_autocmd("VimEnter", {
+            callback = function(data)
+              if is_directory(data.file) then
+                vim.schedule(function()
+                  require("oil").open(data.file)
+                end)
+              end
+            end,
+          })
 
-        -- Handle `:e somedir/` and `gf somedir`
-        vim.api.nvim_create_autocmd("BufEnter", {
-          callback = function()
-            local bufname = vim.api.nvim_buf_get_name(0)
-            if is_directory(bufname) then
-              require("lz.n").trigger_load("oil.nvim")
-            end
-          end,
-        })
-      '';
+          -- Handle `:e somedir/` and `gf somedir`
+          vim.api.nvim_create_autocmd("BufEnter", {
+            callback = function()
+              local bufname = vim.api.nvim_buf_get_name(0)
+              if is_directory(bufname) then
+                require("lz.n").trigger_load("oil.nvim")
+              end
+            end,
+          })
+        '';
   };
 }