--- a/config/plugins/render-markdown.nix
+++ b/config/plugins/render-markdown.nix
@@ -1,10 +1,12 @@
-_: {
+{ config, ... }:
+{
   config = {
     # https://nix-community.github.io/nixvim/plugins/render-markdown
     # https://github.com/MeanderingProgrammer/render-markdown.nvim
     plugins.render-markdown = {
       enable = true;
       lazyLoad.settings.ft = "markdown";
+      lazyLoad.enable = config.lazyLoad.enable;
       settings = {
         sign.enabled = false;
         heading.border = true;