--- a/config/plugins/vim-matchup.nix
+++ b/config/plugins/vim-matchup.nix
@@ -4,11 +4,11 @@ _: {
   # https://nix-community.github.io/nixvim/plugins/vim-matchup
   config = {
     plugins.vim-matchup = {
+      enable = true;
       treesitter = {
         enable = true;
         include_match_words = true;
       };
-      enable = true;
     };
   };
 }