BEFORE:
```nu
> (1..100 | each {|| timeit { nvim -c ":q" } } | math sum) / 100
197ms 887µs 118ns
```
AFTER:
```nu
> (1..100 | each {|| timeit { nvim -c ":q" } } | math sum) / 100
102ms 19µs 355ns
```

Startup time saved on average: 95ms 867µs 763ns